requirements.txt
appnope==0.1.0
decorator==4.0.10
Django==1.10
django-filter==0.14.0
django-graphiql==0.4.4
django-treebeard==4.0.1
gnureadline==6.3.3
graphene==0.10.2
graphql-core==0.5.3
graphql-django-view==1.3
graphql-relay==0.4.4
ipython==4.2.1
ipython-genutils==0.1.0
iso8601==0.1.11
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.3.1
promise==0.4.2
ptyprocess==0.5.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
traitlets==4.2.2
psycopg2
gunicorn
